Backend/translatorapp_v2/requirements/base.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.8.30
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pydantic==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.23.2
pytest==8.3.2
pytest-django==4.9.0
pytest-mock==3.14.0
python-dotenv==1.0.1
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
requests==2.32.3
